fix coor bug

+++ core/data/batch_creator.py
@@ -57,7 +57,7 @@ class BatchCreator(data.Dataset):
             with h5py.File(input_data, 'r') as raw:
                 self.input_data.append(torch.from_numpy((raw['image'][()].astype(np.float32)-128) / 33.0))
                 self.label_data.append(torch.from_numpy(raw['label'][()]))
-                self.coor.append(raw['coor'][()].astype(np.uint8))
+                self.coor.append(raw['coor'][()])
                 self.seed.append(logit(np.full(list(raw['label'][()].shape), 0.05, dtype=np.float32)))
 
     def __getitem__(self, idx):
